Retrieve last item from Firebase.Rides and put it in Ride object

+++ app/src/main/java/com/example/caspe/kilonotes/model/Ride.java
@@ -8,7 +8,9 @@ import com.google.firebase.database.Exclude;
 
 public class Ride {
     public String userName; // Todo: check if firebase auth uses string or int id's
-    public int startDistance;
-    public int endDistance;
+    public long startDistance;
+    public long endDistance;
     public String date;
+
+
 }
